% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdseq_result_pbmcmix-data.R
\docType{data}
\name{cdseq.result}
\alias{cdseq.result}
\title{Output of synthetic mixtures of PBMC scRNAseq data}
\format{
numeric matrix
}
\source{
This is the CDSeq output of synthetic PBMC mixtures
}
\usage{
data(SyntheticMixtureData)
}
\description{
Output of synthetic mixtures of PBMC scRNAseq data
}
\author{
Kai Kang
}
