% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell2RNA.R
\name{Cell2RNA}
\alias{Cell2RNA}
\title{Cell proportion to RNA proportion
\code{Cell2RNA} converts Cell proportion to RNA proportion}
\usage{
Cell2RNA(eta, cellprop)
}
\arguments{
\item{eta}{numeric vector represents the different amounts of RNA produced by different cell types}

\item{cellprop}{sample-specific cell-type proportion}
}
\value{
Cell2RNA returns sample-specific cell-type RNA proportion
}
\description{
Cell proportion to RNA proportion
\code{Cell2RNA} converts Cell proportion to RNA proportion
}
