% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeAssign.R
\name{cellTypeAssign}
\alias{cellTypeAssign}
\title{Assign cell types using correlation matrix computed using cell-type-specific GEPs and reference GEPs.
\code{cellTypeAssign} assigns CDSeq-identified cell types to reference profile.}
\usage{
cellTypeAssign(corMat, threshold = 0.8)
}
\arguments{
\item{corMat}{correlation matrix between CDSeq-estimated GEPs and reference GEPs.}

\item{threshold}{only the correlations that are above threshold will be considered.}
}
\value{
cellTypeAssign returns a vector of cell type assignment to the referecen profile.
}
\description{
Assign cell types using correlation matrix computed using cell-type-specific GEPs and reference GEPs.
\code{cellTypeAssign} assigns CDSeq-identified cell types to reference profile.
}
