% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2rpkm.R
\name{gene2rpkm}
\alias{gene2rpkm}
\title{gene2rpkm outputs the rpkm normalizations of the CDSeq-estimated GEPs.
\code{gene2rpkm} outputs the rpkm normalizations of the CDSeq-estimated GEPs.}
\usage{
gene2rpkm(gene_rate, gene_effective_length, cell_line_counts)
}
\arguments{
\item{gene_rate}{CDSeq-estimated GEP normalized by gene length.}

\item{gene_effective_length}{gene effective length which is the gene length minus the read length plus one.}

\item{cell_line_counts}{RNA-Seq read counts data of cell lines}
}
\value{
gene2rpkm returns rpkm normalization of the CDSeq-estimated GEPs.
}
\description{
gene2rpkm outputs the rpkm normalizations of the CDSeq-estimated GEPs.
\code{gene2rpkm} outputs the rpkm normalizations of the CDSeq-estimated GEPs.
}
