% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hungarian_Rcpp}
\alias{hungarian_Rcpp}
\title{This is the Hungarian algorithm wrapper for cell type assignment
\code{hungarian_Rcpp} returns cell type assignment given reference GEPs}
\usage{
hungarian_Rcpp(costMat)
}
\arguments{
\item{costMat}{correlation matrix}
}
\value{
cost for the assignment and cell type assignment
}
\description{
This is the Hungarian algorithm wrapper for cell type assignment
\code{hungarian_Rcpp} returns cell type assignment given reference GEPs
}
