% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{intersection}
\alias{intersection}
\title{\code{intersection} take intersection of multiple lists and return the common set and index}
\usage{
intersection(list.vector, order = "sort")
}
\arguments{
\item{list.vector}{this is a list of list contain all the data.}

\item{order}{this is either sort or stable. If choose sort, the output common value will be sorted. If choose stable, the output common value will be in the same order as appear in the first element in list.vector.}
}
\value{
The common values among lists and their indices.
intersection function: 
input: list.vector is a list of list contain all the data
for example, if we need to find the common elements of a, b, c, then input should be list(a,b,c)
}
\description{
\code{intersection} take intersection of multiple lists and return the common set and index
}
