% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{loglikelihood computes the log likelihood of the CDSeq model.
\code{loglikelihood} outputs the value of log likelihood}
\usage{
loglikelihood(estProp, estGEP, mydata, alpha, beta)
}
\arguments{
\item{estProp}{CDSeq-estimated cell type proportions.}

\item{estGEP}{CDSeq-estimated cell-type-specific GEPs.}

\item{mydata}{input bulk RNA-seq data.}

\item{alpha}{hyperparameter for cell type proportion estimation.}

\item{beta}{hyperparameter for cell-type-specific GEP estimation.}
}
\value{
loglikelihood returns log likelihood values.
}
\description{
loglikelihood computes the log likelihood of the CDSeq model.
\code{loglikelihood} outputs the value of log likelihood
}
