% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_rep.R
\name{max_rep}
\alias{max_rep}
\title{\code{max_rep} Find the element that repeats the most in a given vector and calculate its proportion.}
\usage{
max_rep(v)
}
\arguments{
\item{v}{a vector}
}
\value{
max_rep_value contains two elements: max_element and max_element_proportion. max_element is the element that repeats the most in v, and max_element_proportion
is its proportion.
}
\description{
\code{max_rep} Find the element that repeats the most in a given vector and calculate its proportion.
}
