% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_CDSeq_estprop-data.R
\docType{data}
\name{merge_df}
\alias{merge_df}
\title{Data frame for keeping the CDSeq-estimated cell type proportions for PBMC mixtures}
\format{
data frame
}
\source{
CDSeq estimated cell type proportions for cell type number 3, 6, 9 and 12
}
\usage{
data(SyntheticMixtureData)
}
\description{
Data frame for keeping the CDSeq-estimated cell type proportions for PBMC mixtures
}
\author{
Kai Kang
}
