% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_ggplot-data.R
\docType{data}
\name{pbmc_ggplot}
\alias{pbmc_ggplot}
\title{ggplot figures of comparison between CDSeq-estimated cell type proportion and ground truth}
\format{
ggplot object
}
\source{
CDseq-estimated cell type proportion and ground truth
}
\usage{
data(SyntheticMixtureData)
}
\description{
ggplot figures of comparison between CDSeq-estimated cell type proportion and ground truth
}
\author{
Kai Kang
}
