% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_mix-data.R
\docType{data}
\name{pbmc_mix}
\alias{pbmc_mix}
\title{Synthetic bulk RNA-seq read counts data of PBMC single cell data}
\format{
numeric matrix
}
\source{
we downloaded the PBMC scRNAseq and generated the mixing proportions randomly and produced the mixtures
}
\usage{
data(SyntheticMixtureData)
}
\description{
Synthetic bulk RNA-seq read counts data of PBMC single cell data
}
\author{
Kai Kang
}
