% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refGEP-data.R
\docType{data}
\name{refGEP}
\alias{refGEP}
\title{GEPs of six component pure cell lines}
\format{numeric matrix}
\usage{
data(SyntheticMixtureData)
}
\description{
GEPs of six component pure cell lines
}
\author{
Kai Kang
}
