% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdseq_result2-data.R
\docType{data}
\name{result2}
\alias{result2}
\title{CDSeq result of synthetic bulk RNA-seq read counts data of six cell types}
\format{
numeric matrix
}
\source{
CDSeq estimates of mixtureGEP
}
\usage{
data(SyntheticMixtureData)
}
\description{
CDSeq result of synthetic bulk RNA-seq read counts data of six cell types
}
\author{
Kai Kang
}
