% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA2Cell.R
\name{RNA2Cell}
\alias{RNA2Cell}
\title{RNA proportion to cell proportion
\code{RNA2Cell} converts RNA proportion to cell proportion}
\usage{
RNA2Cell(eta, rnaprop)
}
\arguments{
\item{eta}{numeric vector represents the different amounts of RNA produced by different cell types}

\item{rnaprop}{sample-specific cell-type RNA proportion}
}
\value{
RNA2Cell returns sample-specific cell-type proportion
}
\description{
RNA proportion to cell proportion
\code{RNA2Cell} converts RNA proportion to cell proportion
}
