% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_annotation-data.R
\docType{data}
\name{sc_annotation}
\alias{sc_annotation}
\title{Cell type annotation of the PBMC single cell data}
\format{
numeric matrix
}
\source{
We used the annotation provided by the single cell data
}
\usage{
data(SyntheticMixtureData)
}
\description{
Cell type annotation of the PBMC single cell data
}
\author{
Kai Kang
}
