% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_gep-data.R
\docType{data}
\name{sc_gep}
\alias{sc_gep}
\title{PBMC single cell RNAseq read counts that used for creating synthetic PBMC mixtures}
\format{
numeric matrix
}
\source{
we downloaded the PBMC RNA-seq data and generated the mixing proportions randomly and produced the mixtures
}
\usage{
data(SyntheticMixtureData)
}
\description{
PBMC single cell RNAseq read counts that used for creating synthetic PBMC mixtures
}
\author{
Kai Kang
}
