% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seedMT}
\alias{seedMT}
\title{This is the Mersenne Twister random number generator.
\code{cokus} generates pseudorandom integers uniformly distributed in 0..(2^32 - 1).}
\usage{
seedMT(seed)
}
\arguments{
\item{seed}{odd number for seeding}
}
\value{
random integers  uniformly distributed in 0..(2^32 - 1).
}
\description{
This is the Mersenne Twister random number generator.
\code{cokus} generates pseudorandom integers uniformly distributed in 0..(2^32 - 1).
}
\author{
Shawn Cokus (Cokus@math.washington.edu)
}
