% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SyntheticMixtureData}
\alias{SyntheticMixtureData}
\title{Synthetic bulk RNA-seq read counts data of six cell types, PBMC mixtures using scRNASeq and some preliminary results}
\format{
A matrix of read counts data containing 40 synthetic mixtures with 500 genes
}
\source{
we downloaded the pure cell line RNA-seq data and generated the mixing proportions randomly and produced the mixtures
}
\usage{
data(SyntheticMixtureData)
}
\description{
Synthetic bulk RNA-seq read counts data of six cell types, PBMC mixtures using scRNASeq and some preliminary results
}
\examples{
data(SyntheticMixtureData)
}
\author{
Kai Kang
}
