% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueGEPgene-data.R
\docType{data}
\name{true_GEP_gene}
\alias{true_GEP_gene}
\title{True GEPs of the six component cell types normalized by gene length}
\format{
numeric matrix
}
\usage{
data(SyntheticMixtureData)
}
\description{
True GEPs of the six component cell types normalized by gene length
}
\author{
Kai Kang
}
