% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueGEPread-data.R
\docType{data}
\name{true_GEP_read}
\alias{true_GEP_read}
\title{True GEPs of the six component cell types unnormalized by gene length}
\format{numeric matrix}
\usage{
data(SyntheticMixtureData)
}
\description{
True GEPs of the six component cell types unnormalized by gene length
}
\author{
Kai Kang
}
