% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueGEPrpkm-data.R
\docType{data}
\name{true_GEP_rpkm}
\alias{true_GEP_rpkm}
\title{True GEPs of the six component cell types RPKM normalization}
\format{numeric matrix}
\usage{
data(SyntheticMixtureData)
}
\description{
True GEPs of the six component cell types RPKM normalization
}
\author{
Kai Kang
}
