% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_true_prop-data.R
\docType{data}
\name{true_prop}
\alias{true_prop}
\title{True cell type proportion in the PBMC synthetic mixtures}
\format{
numeric matrix
}
\source{
randomly generated
}
\usage{
data(SyntheticMixtureData)
}
\description{
True cell type proportion in the PBMC synthetic mixtures
}
\author{
Kai Kang
}
