% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeProp-data.R
\docType{data}
\name{true_prop_cell}
\alias{true_prop_cell}
\title{True cell proportions of the mixtures}
\format{numeric matrix}
\source{
cell type proportions are randomly generated
}
\usage{
data(SyntheticMixtureData)
}
\description{
True cell proportions of the mixtures
}
\author{
Kai Kang
}
