% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeRNAProp-data.R
\docType{data}
\name{true_prop_RNA}
\alias{true_prop_RNA}
\title{True cell type RNA proportions}
\format{
numeric matrix
}
\source{
randomly generated
}
\usage{
data(SyntheticMixtureData)
}
\description{
True cell type RNA proportions
}
\author{
Kai Kang
}
