% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SyntheticMixtureData}
\alias{SyntheticMixtureData}
\title{Synthetic bulk RNA-seq read counts data of six cell types}
\format{This example dataset contains the following}
\source{
we downloaded the pure cell line RNA-seq data and generated the mixing proportions randomly and produced the mixtures
}
\usage{
data(SyntheticMixtureData)
}
\description{
Synthetic bulk RNA-seq read counts data of six cell types
}
\author{
Kai Kang
}
